FIXED BRICKS

+++ Stage 2/RASTER.C
@@ -82,7 +82,7 @@ void plot_ball(UINT8 *base, int x, int y, UINT16 bitmap[8]) {
     }
 }
 
-void plot_brick (UINT8 *base, int x, int y, UINT8 bitmap[8]) {
+void plot_brick (UINT8 *base, int x, int y, UINT16 bitmap[8]) {
 	int row;
 	for (row = 0; row < 8; row++) {  /* iterate through rows */
 		UINT16 current_row = bitmap[row];
